fix duplicate push back
