+++ Inc/ST-LIB_HIGH/Protections/ProtectionManager.hpp
@@ -22,3 +22,17 @@ public:
 
     static void check_protections();
 };
+
+template<class... AdditionalProtections>
+Protection& ProtectionManager::_add_protection(double* src, ProtectionType protection_type,double bound, AdditionalProtections... more_protections) {
+    protections.push_back(Protection(src));
+    protections.back().add_protection(protection_type, bound, more_protections...);
+    return protections.back();
+}
+
+template<class... AdditionalProtections>
+Protection& ProtectionManager::_add_protection(double* src, ProtectionType protection_type, double lower_bound, double upper_bound, AdditionalProtections... more_protections) {
+    protections.push_back(Protection(src));
+    protections.back().add_protection(protection_type, lower_bound, upper_bound, more_protections...);
+    return protections.back();
+}
