+++ Inc/ST-LIB_HIGH/ST-LIB_HIGH.hpp
@@ -6,3 +6,6 @@
  */
 
 #pragma once
+
+#include "Protections/Protection.hpp"
+#include "Protections/ProtectionManager.hpp"
